move database connection and models into database/
