+++ src-tauri/src/database/mod.rs
@@ -1,0 +1,2 @@
+pub mod connection;
+pub mod models;
