+++ src-tauri/src/database/models.rs
@@ -1,5 +1,3 @@
-// Generated by diesel_ext
-
 #![allow(unused)]
 #![allow(clippy::all)]
